Add BOMEX single stack
