+++ docs/Project.toml
@@ -1,4 +1,5 @@
 [deps]
+ArgParse = "c7e460c6-2fb9-53a9-8c5b-16f535851c63"
 CLIMAParameters = "6eacf6c3-8458-43b9-ae03-caf5306d3d53"
 ClimateMachine = "777c4786-024f-11e9-21a3-85d5d4106250"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
