+++ src/Common/TurbulenceConvection/boundary_conditions.jl
@@ -2,6 +2,7 @@
 
 export TurbConvBC,
     NoTurbConvBC,
+    turbconv_bcs,
     turbconv_boundary_state!,
     turbconv_normal_boundary_flux_second_order!
 
@@ -13,6 +14,8 @@ abstract type TurbConvBC end
 Boundary conditions are not applied
 """
 struct NoTurbConvBC <: TurbConvBC end
+
+turbconv_bcs(::NoTurbConv) = NoTurbConvBC()
 
 function turbconv_boundary_state!(nf, bc_turbulence::NoTurbConvBC, bl, args...)
     nothing
